replaced cuda_hybrid file
